refactor!: move batch methods to private mixin to privatize methods
